+++ src/Routes/employee.js
@@ -1,10 +1,11 @@
 import React from 'react';
 import { Switch, Route, useRouteMatch, Redirect } from 'react-router-dom';
-import Home from 'Components/Employee/Home';
-import Timesheet from 'Components/Employee/Timesheet';
-import Projects from 'Components/Employee/Projects';
 import Layout from 'Components/Layout';
+import Home from 'Components/Employee/Home';
 import Profile from 'Components/Employee/Profile';
+import Projects from 'Components/Employee/Projects';
+import Timesheets from 'Components/Employee/Timesheet';
+import Tasks from 'Components/Employee/Timesheet/Tasks';
 import { faClock, faFolderClosed, faHouse } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
@@ -24,9 +25,10 @@ const Employee = () => {
   return (
     <Layout routes={employeesRoutes} logout>
       <Switch>
-        <Route exact path={`${url}/time-sheet`} component={Timesheet} />
-        <Route exact path={`${url}/projects`} component={Projects} />
         <Route exact path={`${url}/profile`} component={Profile} />
+        <Route exact path={`${url}/projects`} component={Projects} />
+        <Route exact path={`${url}/time-sheet`} component={Timesheets} />
+        <Route exact path={`${url}/time-sheet/:id`} component={Tasks} />
         <Route exact path={`${url}/`} component={Home} />
         <Redirect to={`${url}`} />
       </Switch>
